requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
mws==0.8.6
openpyxl==2.6.3
pyaml==19.4.1
PyQt5==5.13.0
PyQt5-sip==4.19.18
PyYAML==5.1.2
requests==2.22.0
schedule==0.6.0
soupsieve==1.9.3
urllib3==1.25.3